services/monolith_backup/backend/requirements.txt

# Core Django
Django>=4.2,<5.0
psycopg2-binary>=2.9 
python-dotenv>=1.0

djangorestframework>=3.14
django-cors-headers>=4.0
channels>=4.0
asgiref>=3.7
channels-redis>=4.1

djangorestframework-simplejwt>=5.3
social-auth-app-django>=5.4
django-allauth
dj-rest-auth
django-storages>=1.14
boto3>=1.34
Pillow>=10.0
moviepy>=1.0.3

celery>=5.3
redis>=5.0

transformers>=4.41
torch>=2.3
gunicorn>=21.2
whitenoise>=6.6
requests>=2.31
uuid>=1.30
python-magic>=0.4.27
ffmpeg-python>=0.2.0
python-decouple>=3.8
ipython>=8.24
django-debug-toolbar>=4.3
pytest-django>=4.7
